% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{class}
\name{AbstractRamlAlgObject-class}
\alias{AbstractRamlAlgObject-class}
\title{An abstract class that represents anything that conforms to algebraic rules.}
\description{
An abstract class that represents anything that conforms to algebraic rules.
}
