% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{class}
\name{AffineExpr-class}
\alias{AffineExpr-class}
\title{A class that represents an affine combination of variables.}
\description{
A class that represents an affine combination of variables.
}
\section{Slots}{

 \describe{
   \item{\code{vars}:}{Object of class \code{"character"}-- the variables included in this affine expression}
   \item{\code{coefs}:}{Object of class \code{"numeric"}, the scalar coefficients matching the the \code{vars}}
   \item{\code{offset}:}{Object of class \code{"numeric"}, which indicates the scalar that offsets this affine expression from a linear expression.}
 }
}

