% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{dot}
\alias{dot}
\alias{dot,numeric,ramlArray-method}
\alias{dot,ramlArray,numeric-method}
\title{Takes the inner product of a variable array and a numeric array.}
\usage{
dot(a, b)

\S4method{dot}{numeric,ramlArray}(a, b)

\S4method{dot}{ramlArray,numeric}(a, b)
}
\arguments{
\item{a}{An array of class \code{"numeric"}.}

\item{b}{An array of variables.}
}
\description{
Takes the inner product of a variable array and a numeric array.
}
\examples{
m <- Model()
m$var(x[1:3] >= 0)
m$constraint(dot(x, c(1,1,1)) >= 1)
show(m)
}
