% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{dual}
\alias{dual}
\alias{dual,RAMLModel,ramlComparison-method}
\title{Extracts the dual value (shadow cost) of a constraint.}
\usage{
dual(m, constr)

dual(m, constr)
}
\arguments{
\item{m}{The model object.}

\item{constr}{The constraint in question.}
}
\value{
The dual value (shadow cost) of a constraint. Returns \code{NA} if there are integer or binary variables in the model.
}
\description{
Extracts the dual value (shadow cost) of a constraint.
}
\examples{
m <- Model()
m$var(x >= 0)
m$objective(x)
m$constraint(x >= 0.1)
m$solve()
dual(m, x >= 0.1)  # 1.0
}
