% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{Model}
\alias{Model}
\title{Create an optimization model.}
\usage{
Model()
}
\description{
Creates a object that will hold your optimization model.
 You can use the methods of this object to define your optimization problem.
}
\examples{
m <- Model()
m$var(x >= 0)
m$var(y >= 0)
m$constraint(x + y <= 2)
m$sense <- "max"
m$objective(2*x + y)
m$solve()
value(x)
}
