% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{+,AffineExpr,AffineExpr-method}
\alias{+,AffineExpr,AffineExpr-method}
\title{Adds two affine expressions together using basic algebra.}
\usage{
\S4method{+}{AffineExpr,AffineExpr}(e1, e2)
}
\arguments{
\item{e1}{The first AffineExpr to be added.}

\item{e2}{The second AffineExpr}
}
\description{
Adds two affine expressions together using basic algebra.
}
