% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{+,ramlVariable,ramlVariable-method}
\alias{+,ramlVariable,ramlVariable-method}
\title{Adds two variables together to form an affine expression.}
\usage{
\S4method{+}{ramlVariable,ramlVariable}(e1, e2)
}
\arguments{
\item{e1}{The first variable to be added.}

\item{e2}{The second variable to be added.}
}
\description{
Adds two variables together to form an affine expression.
}
