% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{print.ramlArray}
\alias{print.ramlArray}
\title{Nice display for arrays.}
\usage{
\method{print}{ramlArray}(x, ...)
}
\arguments{
\item{x}{The variable to be printed, of the "ramlArray" class.}

\item{...}{Other arguments to pass on to print.}
}
\description{
Nice display for arrays.
}
