% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{print.ramlVariable}
\alias{print.ramlVariable}
\title{Nice display for single variables.}
\usage{
\method{print}{ramlVariable}(x, ...)
}
\arguments{
\item{x}{The variable to be printed, of the "ramlVariable" class.}

\item{...}{Other arguments to pass on to print.}
}
\description{
Nice display for single variables.
}
