% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{+,ramlVariable,numeric-method}
\alias{+,ramlVariable,numeric-method}
\alias{+,numeric,ramlVariable-method}
\alias{+,ramlVariable,ramlVariable-method}
\alias{+,ramlVariable,AffineExpr-method}
\alias{+,AffineExpr,numeric-method}
\alias{+,numeric,AffineExpr-method}
\alias{+,AffineExpr,ramlVariable-method}
\alias{+,AffineExpr,AffineExpr-method}
\alias{*,ramlVariable,numeric-method}
\alias{*,AffineExpr,numeric-method}
\alias{*,numeric,AbstractRamlAlgObject-method}
\alias{-,AbstractRamlAlgObject,numeric-method}
\alias{-,numeric,AbstractRamlAlgObject-method}
\alias{-,AbstractRamlAlgObject,AbstractRamlAlgObject-method}
\alias{/,AbstractRamlAlgObject,numeric-method}
\title{Algebra within the raml ecosystem.}
\usage{
foo

\S4method{+}{numeric,ramlVariable}(e1, e2)

\S4method{+}{ramlVariable,ramlVariable}(e1, e2)

\S4method{+}{ramlVariable,AffineExpr}(e1, e2)

\S4method{+}{AffineExpr,numeric}(e1, e2)

\S4method{+}{numeric,AffineExpr}(e1, e2)

\S4method{+}{AffineExpr,ramlVariable}(e1, e2)

\S4method{+}{AffineExpr,AffineExpr}(e1, e2)

\S4method{*}{ramlVariable,numeric}(e1, e2)

\S4method{*}{AffineExpr,numeric}(e1, e2)

\S4method{*}{numeric,AbstractRamlAlgObject}(e1, e2)

\S4method{-}{AbstractRamlAlgObject,numeric}(e1, e2)

\S4method{-}{numeric,AbstractRamlAlgObject}(e1, e2)

\S4method{-}{AbstractRamlAlgObject,AbstractRamlAlgObject}(e1, e2)

\S4method{/}{AbstractRamlAlgObject,numeric}(e1, e2)
}
\arguments{
\item{e1}{The first algebraic object.}

\item{e2}{The second algebraic object.}
}
\description{
Algebra within the raml ecosystem.
}
