% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\docType{class}
\name{RAML}
\alias{RAML}
\alias{Model}
\title{An object containing all data necessary for preference elicitation.}
\description{
An object containing all data necessary for preference elicitation.
}
\section{Fields}{

\describe{
\item{\code{data}}{A matrix or dataframe of data.}
}}


