% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{class}
\name{ramlComparison-class}
\alias{ramlComparison-class}
\title{A class that represents an affine combination of variables.}
\description{
A class that represents an affine combination of variables.
}
\section{Slots}{

 \describe{
   \item{\code{lhs}:}{Object of class \code{"AffineExpr"}, the left-hand side of the comparison.}
   \item{\code{rhs}:}{Object of class \code{"AffineExpr"}, the rightt-hand side of the comparison.}
   \item{\code{comp}:}{Object of class \code{"character"}, which indicates if the Comparison being made is \code{">="}, \code{"<="}, or \code{"=="}.}
 }
}

