% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{class}
\name{RAMLModel}
\alias{RAMLModel}
\alias{.Model}
\title{A class that holds all the necessary data to build a linear programming model.}
\description{
A class that holds all the necessary data to build a linear programming model.
}
\section{Fields}{

\describe{
\item{\code{sense}}{Should the objective be minimized ("min") or maximized ("max")?}
}}


