% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{class}
\name{ramlVariable-class}
\alias{ramlVariable-class}
\title{A class that represents a variable}
\description{
A class that represents a variable
}
\section{Slots}{

 \describe{
   \item{\code{name}:}{Objectx of class \code{"character"}-- the name of the variable in the user's scope.}
   \item{\code{bounds}:}{Object of class \code{"numeric"}, which defines the lower and upper bounds of the variable's domain.}
   \item{\code{integer}:}{Object of class \code{"character"}, which defines if the variable is a \code{"Real"}, \code{"Integer"}, or \code{"Binary"}}
   \item{\code{value}:}{Object of class \code{"numeric"}, the value of the variable at the model's optimum.}
 }
}

