% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{rsum}
\alias{rsum}
\title{Helper function to take sums over variable arrays.}
\usage{
rsum(var, ...)
}
\arguments{
\item{var}{The indexed variable to be summed.}

\item{...}{The indicies to iterate through.}
}
\description{
Helper function to take sums over variable arrays.
}
\examples{
m <- Model()
m$var(x[1:3] >= 0)
rsum(x[i], i = 1:3)
m$var(y[1:2, 1:2] >= 0)
rsum(y[i,j], i = 1:2, j = 1:2)
}
