% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{+,AffineExpr,numeric-method}
\alias{+,AffineExpr,numeric-method}
\alias{+,numeric,AffineExpr-method}
\title{Adds a scalar to an affine expression.}
\usage{
\S4method{+}{AffineExpr,numeric}(e1, e2)

\S4method{+}{numeric,AffineExpr}(e1, e2)
}
\arguments{
\item{e1}{The AffineExpr to be added.}

\item{e2}{The scalar to be added to the affine expression.}
}
\description{
Adds a scalar to an affine expression.

Adds a scalar to an affine expression.
}
