% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{show,AffineExpr-method}
\alias{show,AffineExpr-method}
\title{Nice display for affine expressions.}
\usage{
\S4method{show}{AffineExpr}(object)
}
\arguments{
\item{object}{The expression to be printed, of the "AffineExpr" class.}
}
\description{
Nice display for affine expressions.
}
