% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{show,ramlArray-method}
\alias{show,ramlArray-method}
\title{Nice display for arrays.}
\usage{
\S4method{show}{ramlArray}(object)
}
\arguments{
\item{object}{The variable to be printed, of the "ramlArray" class.}
}
\description{
Nice display for arrays.
}
