% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{show,ramlComparison-method}
\alias{show,ramlComparison-method}
\title{Nice display for algabraic comparisons.}
\usage{
\S4method{show}{ramlComparison}(object)
}
\arguments{
\item{object}{The Comparison expression to be displayed.}
}
\description{
Nice display for algabraic comparisons.
}
