% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{show,ramlVariable-method}
\alias{show,ramlVariable-method}
\title{Nice display for single variables.}
\usage{
\S4method{show}{ramlVariable}(object)
}
\arguments{
\item{object}{The variable to be printed, of the "ramlVariable" class.}

\item{...}{Other arguments to pass on to print.}
}
\description{
Nice display for single variables.
}
