% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{value}
\alias{value}
\alias{value,ramlAlgObject-method}
\title{A helper function that extracts the value of a variable.}
\usage{
value(x)

\S4method{value}{ramlAlgObject}(x)
}
\arguments{
\item{x}{The object.}
}
\description{
A helper function that extracts the value of a variable.
}
