% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{+,ramlVariable,AffineExpr-method}
\alias{+,ramlVariable,AffineExpr-method}
\alias{+,AffineExpr,ramlVariable-method}
\title{Adds a variable to an affine expression.}
\usage{
\S4method{+}{ramlVariable,AffineExpr}(e1, e2)

\S4method{+}{AffineExpr,ramlVariable}(e1, e2)
}
\arguments{
\item{e1}{The AffineExpr to be added.}

\item{e2}{The variable to be added to the affine expression.}
}
\description{
Adds a variable to an affine expression.
}
