% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{+,ramlVariable,numeric-method}
\alias{+,ramlVariable,numeric-method}
\alias{+,numeric,ramlVariable-method}
\title{Adds a scalar to a variable, resulting in an affine expression.}
\usage{
\S4method{+}{ramlVariable,numeric}(e1, e2)

\S4method{+}{numeric,ramlVariable}(e1, e2)
}
\arguments{
\item{e1}{The first variable to be added.}

\item{e2}{The second variable to be added.}
}
\description{
Adds a scalar to a variable, resulting in an affine expression.
}
