% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{dual}
\alias{dual}
\alias{dual,RAMLModel,ramlComparison-method}
\title{A helper function that extracts the dual value (shadow cost) of a constraint.}
\usage{
dual(m, constr)

\S4method{dual}{RAMLModel,ramlComparison}(m, constr)
}
\arguments{
\item{m}{The model object.}

\item{constr}{The constraint in question.}
}
\value{
The dual value (shadow cost) of a constraint. Returns \code{NA} if there are integer or binary variables in the model.
}
\description{
A helper function that extracts the dual value (shadow cost) of a constraint.
}
