% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{[,ramlArray-method}
\alias{[,ramlArray-method}
\alias{+,ramlVariable,numeric-method}
\alias{+,numeric,ramlVariable-method}
\alias{+,ramlVariable,ramlVariable-method}
\alias{+,ramlVariable,AffineExpr-method}
\alias{+,AffineExpr,numeric-method}
\alias{+,numeric,AffineExpr-method}
\alias{+,AffineExpr,ramlVariable-method}
\alias{+,AffineExpr,AffineExpr-method}
\alias{*,ramlVariable,numeric-method}
\alias{*,AffineExpr,numeric-method}
\alias{*,numeric,AbstractRamlAlgObject-method}
\alias{-,AbstractRamlAlgObject,numeric-method}
\alias{-,numeric,AbstractRamlAlgObject-method}
\alias{-,AbstractRamlAlgObject,AbstractRamlAlgObject-method}
\alias{/,AbstractRamlAlgObject,numeric-method}
\alias{>=,ANY,AbstractRamlAlgObject-method}
\alias{<=,ANY,AbstractRamlAlgObject-method}
\alias{==,ANY,AbstractRamlAlgObject-method}
\alias{>=,AbstractRamlAlgObject,ANY-method}
\alias{<=,AbstractRamlAlgObject,ANY-method}
\alias{==,AbstractRamlAlgObject,ANY-method}
\title{Natural subsetting for indicies}
\usage{
\S4method{[}{ramlArray}(x, i, j, ..., drop = TRUE)

Algebra

\S4method{+}{numeric,ramlVariable}(e1, e2)

\S4method{+}{ramlVariable,ramlVariable}(e1, e2)

\S4method{+}{ramlVariable,AffineExpr}(e1, e2)

\S4method{+}{AffineExpr,numeric}(e1, e2)

\S4method{+}{numeric,AffineExpr}(e1, e2)

\S4method{+}{AffineExpr,ramlVariable}(e1, e2)

\S4method{+}{AffineExpr,AffineExpr}(e1, e2)

\S4method{*}{ramlVariable,numeric}(e1, e2)

\S4method{*}{AffineExpr,numeric}(e1, e2)

\S4method{*}{numeric,AbstractRamlAlgObject}(e1, e2)

\S4method{-}{AbstractRamlAlgObject,numeric}(e1, e2)

\S4method{-}{numeric,AbstractRamlAlgObject}(e1, e2)

\S4method{-}{AbstractRamlAlgObject,AbstractRamlAlgObject}(e1, e2)

\S4method{/}{AbstractRamlAlgObject,numeric}(e1, e2)

\S4method{>=}{ANY,AbstractRamlAlgObject}(e1, e2)

\S4method{<=}{ANY,AbstractRamlAlgObject}(e1, e2)

\S4method{==}{ANY,AbstractRamlAlgObject}(e1, e2)

\S4method{>=}{AbstractRamlAlgObject,ANY}(e1, e2)

\S4method{<=}{AbstractRamlAlgObject,ANY}(e1, e2)

\S4method{==}{AbstractRamlAlgObject,ANY}(e1, e2)
}
\arguments{
\item{x}{The object of \code{"ramlArray"} to be subsetted.}

\item{i}{The first index.}

\item{j}{The second index (if needed).}

\item{...}{Additional indicies.}

\item{drop}{(Not used).}

\item{e1}{The first algebraic object.}

\item{e2}{The second algebraic object.}
}
\description{
Natural subsetting for indicies

Algebra within the raml ecosystem.
}
