% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{class}
\name{ramlArray-class}
\alias{ramlArray-class}
\title{A class that represents a variable}
\description{
A class that represents a variable
}
\section{Slots}{

 \describe{
   \item{\code{indicies}:}{Objectx of class \code{"character"}-- a vector of the suffixes appended to this variable for its elements.}
   \item{\code{indexDisplay}:}{Objectx of class \code{"character"}-- human-readable display of these indicies.}
 }
}

