% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{class}
\name{RAMLModel}
\alias{RAMLModel}
\alias{.Model}
\title{A class that holds all the necessary data to build a linear programming model.}
\description{
A class that holds all the necessary data to build a linear programming model.
}
\section{Fields}{

\describe{
\item{\code{sense}}{Should the objective be minimized ("min") or maximized ("max")?}

\item{\code{soln}}{An object of class \code{"OP_solution"} that contains information about the problem solution.}

\item{\code{.__obj}}{(Internal use only) The objective function of the model.}

\item{\code{.__constraints}}{(Internal use only) A list of constraints of the model.}

\item{\code{.__variables}}{(Internal use only) An array of the variables declared in the model.}
}}


