% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{class}
\name{RAMLModel}
\alias{RAMLModel}
\alias{.Model}
\title{An object containing all data necessary for preference elicitation.}
\description{
An object containing all data necessary for preference elicitation.
}
\section{Fields}{

\describe{
\item{\code{sense}}{Should the objective be minimized ("min") or maximized ("max")?}
}}


