% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{class}
\name{ramlVariable-class}
\alias{ramlVariable-class}
\title{A class that represents a single variable}
\description{
A class that represents a single variable
}
\section{Slots}{

 \describe{
   \item{\code{parent}:}{Object of class \code{"character"}-- the name of the array this variable is a part of. Array of length 0 if variable is not part of array..}
 }
}

