% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{methods}
\name{value}
\alias{value}
\alias{value,ramlAlgObject-method}
\title{A helper function that extracts the value of a variable.}
\usage{
value(x)

\S4method{value}{ramlAlgObject}(x)
}
\arguments{
\item{x}{The object.}
}
\description{
A helper function that extracts the value of a variable.
}
\examples{
m <- Model()
m$var(x >= 0)
m$objective(x)
m$constraint(x >= 0.1)
value(x) # NA
m$solve()
value(x) # 0.1
}
